requirements.txt
bson==0.5.10
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
Deprecated==1.2.12
Flask==2.0.1
gcloud==0.18.3
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
jwcrypto==0.9.1
MarkupSafe==2.0.1
oauth2client==4.1.3
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pyparsing==2.4.7
Pyrebase4==4.5.0
python-dateutil==2.8.1
python-dotenv==0.18.0
python-jwt==3.3.0
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.16.0
tzlocal==2.1
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
